requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pgvector
celery
redis
python-multipart
requests
boto3
mistralai
presidio-analyzer
presidio-anonymizer
python-dotenv
pydantic-settings
pydantic[email]
email-validator
PyYAML==6.0.1
reportlab==4.0.7
pdfplumber==0.10.3
google-generativeai
